Only splitting off `flecs-sys` into its own crate. Added `--workspace` to the CI calls. Added a skeleton `README.md` for `flecs-sys`.
