+++ .github/workflows/ci.yml
@@ -16,6 +16,6 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: Build
-        run: cargo build --verbose
+        run: cargo build --workspace --verbose
       - name: Run tests
-        run: cargo test --verbose
+        run: cargo test --workspace --verbose
