+++ flecs-sys/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "flecs-sys"
+description = "Provides low-level bindings to the Entity Component System (ECS) framework flecs"
+version.workspace = true
+edition.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+rust-version.workspace = true
+links = "flecs"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+once_cell = "1.18"
+
+[build-dependencies]
+bindgen = "0.66"
+cc = "1.0.70"
