+++ flecs-sys/README.md
@@ -1,0 +1,1 @@
+# `flecs-sys`
