+++ src/lib.rs
@@ -9,21 +9,7 @@
 
 use std::{any::TypeId, mem::MaybeUninit};
 
-// We generate bindings to an actual source file so that we get better IDE integration
-
-// For now do not export Docs for all the Raw bindings.
-// Sadly to publish on crates.io we cannot write outside the OUT_DIR
-// revisit this later.
-// We will need to expose types that are part of the Rust api at some point
-// #[doc(hidden)]
-// mod bindings;
-// #[doc(hidden)]
-// pub use bindings::*;
-
-mod bindings {
-	include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
-}
-pub use bindings::*;
+pub use flecs_sys::*;
 
 mod binding_util;
 pub use binding_util::*;
@@ -85,21 +71,6 @@ impl AsEcsId for EntityId {
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
-// C Struct initializer Defaults
-//
-impl Default for ecs_entity_desc_t {
-	fn default() -> Self {
-		let desc: ecs_entity_desc_t = unsafe { MaybeUninit::zeroed().assume_init() };
-		desc
-	}
-}
-
-impl Default for ecs_system_desc_t {
-	fn default() -> Self {
-		let desc: ecs_system_desc_t = unsafe { MaybeUninit::zeroed().assume_init() };
-		desc
-	}
-}
 
 // TODO - port more C++ tests to Rust!!!
 //
